refactor(#23): SpectrumView を struct/geometry/rendering の3ファイルに分割
View ファイルは純粋な SwiftUI struct のみを残し、bar ジオメトリを
SpectrumGeometry.swift（純粋・テスト可能）、Canvas 描画を
SpectrumRendering.swift（GraphicsContext 依存・薄い）へ自由関数として分離。
body はそれらを呼ぶだけで計算・描画メソッドを持たない。
